+++ src/application/context/project/init/PlanProjectInitResponse.ts
@@ -1,5 +1,7 @@
+import { AvailableAgent } from "./AgentSelection.js";
 import { PlannedFileChange } from "./PlannedFileChange.js";
 
 export interface PlanProjectInitResponse {
+  readonly availableAgents: readonly AvailableAgent[];
   readonly plannedChanges: PlannedFileChange[];
 }